product_recommendation_api/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv

